bug fixes post kaimerge
- Fixed a couple of cases of `{e}` not referring to a variable (missing Exception as e)
- Fixed a few cases of missing `self.` for variables or a `self` arg in a class method
- Removed a duplicated dictionary entry in `models\v2.py`.
- Fixed adding a team, which currently broken (500 and it doesn't work) due to a missing import, probably missed because it would have required a circular reference (`database\functions.py` logically required `base\team.py` and vice versa).
